refactor(zb_io): split monolothic download functionality into submodules
